refactor: 파일에서 item 데이터를 로드 및 파싱 하는 메서드 클래스로 추출
