+++ src/main/java/store/model/item/Inventory.java
@@ -1,19 +1,15 @@
 package store.model.item;
 
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.List;
 import store.error.FileParsingException;
 import store.error.PromotionConfirmationForFreeException;
 import store.model.order.OrderItem;
 import store.model.promotion.Promotion;
 import store.model.promotion.PromotionCalculation;
 import store.model.promotion.PromotionManager;
-import store.utils.MarkdownReader;
 
 public class Inventory {
     private static final String PRODUCTS_FILE_PATH = "src/main/resources/products.md";
-    private static final String NULL = "null";
 
     private final Items items;
     private final PromotionManager promotionManager;
@@ -24,7 +20,7 @@ public class Inventory {
 
     public Inventory(String productsFilePath) {
         try {
-            this.items = loadItemsFromFile(productsFilePath);
+            this.items = new Items(ItemLoader.getInstance().loadItems(productsFilePath));
             this.promotionManager = new PromotionManager();
             promotionManager.loadPromotions();
         } catch (IOException e) {
@@ -38,48 +34,6 @@ public class Inventory {
                 orderItem.setPrice(item.getPrice());
             }
         }
-    }
-
-    private Items loadItemsFromFile(String productsFilePath) {
-        try {
-            List<String[]> itemData = MarkdownReader.readFile(productsFilePath);
-            List<Item> items = parseItems(itemData);
-            return new Items(items);
-        } catch (IOException e) {
-            throw new FileParsingException(e);
-        }
-    }
-
-    private List<Item> parseItems(List<String[]> itemData) {
-        List<Item> items = new ArrayList<>();
-        for (int i = 0; i < itemData.size(); i++) {
-            Item item = parseItem(itemData.get(i));
-            items.add(item);
-            if (shouldAddNonPromotionItem(item, i, itemData)) {
-                items.add(new Item(item.getName(), item.getPrice(), 0, null));
-            }
-        }
-        return items;
-    }
-
-    private Item parseItem(String[] itemData) {
-        String name = itemData[0];
-        int price = Integer.parseInt(itemData[1]);
-        int quantity = Integer.parseInt(itemData[2]);
-        String promotionName = itemData[3];
-        if (NULL.equals(promotionName)) {
-            promotionName = null;
-        }
-        return new Item(name, price, quantity, promotionName);
-    }
-
-    private boolean shouldAddNonPromotionItem(Item currentItem, int currentIndex, List<String[]> itemData) {
-        if (currentItem.getPromotionName() == null) {
-            return false;
-        }
-        boolean isLastItem = currentIndex == itemData.size() - 1;
-        boolean isDifferentNextItem = !isLastItem && !itemData.get(currentIndex + 1)[0].equals(currentItem.getName());
-        return isLastItem || isDifferentNextItem;
     }
 
     public void consumePromotionItemWithoutPromotion(OrderItem orderItem) {
